# docs/swagger.yaml
definitions:
  domain.BaseUser:
    properties:
      email:
        type: string
      id:
        type: string
      username:
        type: string
    required:
    - email
    - id
    - username
    type: object
...
  domain.FindTeamsResponse:
    properties:
      teams:
        items:
          $ref: '#/definitions/domain.Team'
        type: array
      total:
        type: integer
    required:
    - teams
    - total
    type: object
...
  domain.IDResponse:
    properties:
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        format: uuid
        type: string
    required:
    - id
    type: object
...
  domain.Member:
    properties:
      id:
        type: string
      joined_at:
        type: string
      role:
        $ref: '#/definitions/domain.Role'
      user:
        $ref: '#/definitions/domain.BaseUser'
    required:
    - id
    - joined_at
    - role
    - user
    type: object
  domain.Role:
    enum:
    - admin
    - member
    - owner
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleMember
    - RoleOwner
...
  domain.Team:
    properties:
      created_at:
        type: string
      id:
        type: string
      members:
        items:
          $ref: '#/definitions/domain.Member'
        type: array
      name:
        type: string
      owner:
        $ref: '#/definitions/domain.BaseUser'
      updated_at:
        type: string
    required:
    - created_at
    - id
    - members
    - name
    - owner
    - updated_at
    type: object
  domain.TeamStats:
    properties:
      done_tasks_last_seven_days:
        type: integer
      id:
        type: string
      members_count:
        type: integer
      name:
        type: string
    required:
    - done_tasks_last_seven_days
    - id
    - members_count
    - name
    type: object
  domain.TopUserInTeam:
    properties:
      created_tasks:
        type: integer
      team_id:
        type: string
      team_name:
        type: string
      user:
        $ref: '#/definitions/domain.BaseUser'
    required:
    - created_tasks
    - team_id
    - team_name
    - user
    type: object
  domain.ValidatationErrors:
    properties:
      errors:
        items:
          $ref: '#/definitions/domain.ValidationError'
        type: array
    type: object
  domain.ValidationError:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  httpcommon.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  inputs.CreateTeam:
    properties:
      name:
        example: team 1
        type: string
    required:
    - name
    type: object
...
info:
  contact:
    email: f3d0t0tvmax@yandex.ru
    name: Fedotv Maxim (developer)
  description: Swagger Documentation for mkk_luna_test_rest_api (тестовое задание
    для компании "МКК ЛУНА")
  title: Swagger Documentation for mkk_luna_test_rest_api
  version: "1.0"
paths:
...
  /api/v1/statistics/teams:
    get:
      description: Возвращает статистику по всем командам
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.TeamStats'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
      summary: Статистика по командам
      tags:
      - statistics
  /api/v1/statistics/top-users:
    get:
      description: Возвращает пользователей с наибольшим количеством задач в командах
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.TopUserInTeam'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
      summary: Топ пользователей
      tags:
      - statistics
...
  /api/v1/teams:
    get:
      consumes:
      - application/json
      description: олучить команды пользователя
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.FindTeamsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
      security:
      - BearerAuth: []
      summary: Получить команды пользователя
      tags:
      - teams
    post:
      consumes:
      - application/json
      description: Создать команду
      parameters:
      - description: Объект для создания команды
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/inputs.CreateTeam'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.IDResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ValidatationErrors'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpcommon.MessageResponse'
      security:
      - BearerAuth: []
      summary: Создать команду
      tags:
      - teams
...
securityDefinitions:
  BearerAuth:
    description: Введите "Bearer [ваш JWT токен]"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"